pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "struphy"
version = "2.0.5"
readme = "README.md"
requires-python = ">=3.7, <3.12"
license = {file = "LICENSE"}
authors = [
  { name = "Max Planck Institute for Plasma Physics"},
  { email = "stefan.possanner@ipp.mpg.de"},
  { email = "eric.sonnendruecker@ipp.mpg.de"}
]
description = "Multi-model plasma physics package"
keywords = ["plasma physics, fusion, numerical modeling, partial differential equations, energetic particles"]
classifiers = [
    "Programming Language :: Python :: 3",
    ]
dependencies = [
    'h5py',
    'matplotlib',
    'mpi4py',
    'numpy',
    'pyccel>=1.10.0',
    'PyYAML',
    'scipy',
    'pytest',
    'pytest-mpi',
    'mistune',
    'sphinx',
    'sphinxcontrib-napoleon',
    'pydata-sphinx-theme',
    'nbsphinx',
    'notebook',
    'm2r2',
    'vtk',
    'docutils',
    'wheel',
    'tqdm',
    'gvec-to-python',
    'argcomplete',
    'ipyparallel',
    'pytest-monitor',
]

[project.urls]
homepage = "https://struphy.pages.mpcdf.de/struphy/"
documentation = "https://struphy.pages.mpcdf.de/struphy/"
repository = "https://gitlab.mpcdf.mpg.de/struphy/struphy"
changelog = "https://gitlab.mpcdf.mpg.de/struphy/struphy/-/blob/devel/CHANGELOG.md"
"Bug Tracker" = "https://gitlab.mpcdf.mpg.de/struphy/struphy/-/issues"

[project.scripts]
struphy = "struphy.console.main:struphy"

[tool.setuptools.package-data]
'struphy.fields_background.mhd_equil.eqdsk' = [
            'data/*.high',
        ]

'struphy.fields_background.mhd_equil.gvec' = [
            'ellipstell_v2/newBC_E1D6_M6N6/*.dat',
            'ellipstell_v2/newBC_E1D6_M6N6/*.ini',
            'ellipstell_v2/newBC_E4D6_M6N6/*.dat',
            'ellipstell_v2/newBC_E4D6_M6N6/*.ini',
            'ellipstell_v2/oldBC_E40D5M6N6/*.dat',
            'ellipstell_v2/oldBC_E40D5M6N6/*.ini',
        ]

'struphy.io' = ['batch/*.sh']

'struphy.io.inp' = ['parameters.yml',
                    'tests/*.yml',
                    'tutorials/*.yml',
                    'longer_examples/*.yml',
                    ]
                    
struphy = ['compile_struphy.mk',
           'psydac-0.1.6-py3-none-any.whl',
           ]